Finish 2021 day 10 (all parts)

+++ app/src/main/java/sgtmelon/adventofcode2021_10/SolutionViewModel.kt
@@ -6,5 +6,5 @@ interface SolutionViewModel {
 
     val corruptedPoints: LiveData<Int>
 
-    val incompletePoints: LiveData<Int>
+    val incompletePoints: LiveData<Long>
 }
